adapter - ethereum - move setup fns to `mod test_utils`
